[fix] angle range test
For,
E   TypeError: 'numpy.float64' object cannot be interpreted as an integer

Assign theta_list_linear_len to int for failing ubuntu 20 testcase

For,
E   AssertionError:
E   Arrays are not equal
E   Start of array should be zero-padded
E   Mismatched elements: 102400 / 102400 (100%)
E   Max absolute difference: 518879.41394775
E   Max relative difference: inf
E    x: DataArray([[518879.413948, 518879.413948, 518879.413948, ...,
E               518879.413948, 518879.413948, 518879.413948],
E              [518879.413948, 518879.413948, 518879.413948, ...,...
E    y: array(0)

Update test_with_angle_range, now checks for minimum value padding after interpolation

+++ src/odemis/util/angleres.py
@@ -848,7 +848,7 @@ def project_angular_spectrum_to_grid(
     theta_mn_mx = (-theta_mn_mx[0], theta_mn_mx[1]) if angle_range is None else angle_range
 
     # Define the ideal theta list. Use the min/max angle and linearly spread 1 mrad.
-    theta_list_linear_len = round((theta_mn_mx[1] - theta_mn_mx[0]) / 1e-3)
+    theta_list_linear_len = int(round((theta_mn_mx[1] - theta_mn_mx[0]) / 1e-3))
     # Make sure it's not too small, if range is tiny or image is very fine
     if theta_list_linear_len < data.shape[0]:  # TODO use the length of the biggest theta_list_wl
         logging.info("Increasing theta list length to %s, from %s", data.shape[0], theta_list_linear_len)
